requirements.txt
# Core Framework (Django 6 with built-in tasks)
django>=6.0
uvicorn[standard]>=0.27
gunicorn>=21.2
django-ninja>=1.0
psycopg[binary]>=3.1
django-redis>=5.4

# AI Memory & Vector Storage
lancedb>=0.17.0
pyarrow>=17.0
pandas>=2.2
numpy<2.0.0  # Stable for many ML libraries

# LLM Integration
litellm>=1.81.0
instructor>=1.14
tiktoken>=0.7
langchain-core>=0.1

# Browser & Web Research
playwright>=1.42
beautifulsoup4>=4.12
lxml>=5.1
httpx>=0.27
tavily-python>=0.3
tldextract>=5.1

# File Processing
pypdf>=4.0
pdfplumber>=0.10
pytesseract>=0.3
easyocr>=1.7
Pillow>=10.2
python-docx>=1.1
openpyxl>=3.1
pdf2image>=1.17

# Communication & Voice
python-telegram-bot>=21.0
faster-whisper>=1.0
edge-tts>=6.1
transformers>=4.38
torch>=2.2
soundfile>=0.12
librosa>=0.10

# LLM Providers
ollama>=0.1.7

# Security & Compliance
django-environ>=0.11
libcst>=1.1
RestrictedPython>=7.0
django-ratelimit>=4.1
django-auditlog>=3.4
cryptography>=42.0

# Data & Validation
pydantic>=2.6
orjson>=3.9

# Scheduling
django-crontab>=0.7.1
asgiref>=3.8

# Development & Testing
pytest-django>=4.8
pytest-asyncio>=0.23
black>=24.2
ruff>=0.2
